moved zenodo's upload_type back to shared object; refs #306
